+++ psutil/_common.py
@@ -7,6 +7,7 @@
 """Common objects shared by all _ps* modules."""
 
 from __future__ import division
+import contextlib
 import errno
 import functools
 import os
@@ -138,6 +139,19 @@ def isfile_strict(path):
         return False
     else:
         return stat.S_ISREG(st.st_mode)
+
+
+def supports_ipv6():
+    """Return True if IPv6 is supported on this platform."""
+    if not socket.has_ipv6 or not hasattr(socket, "AF_INET6"):
+        return False
+    try:
+        sock = socket.socket(socket.AF_INET6, socket.SOCK_STREAM)
+        with contextlib.closing(sock):
+            sock.bind(("::1", 0))
+        return True
+    except socket.error:
+        return False
 
 
 def sockfam_to_enum(num):